# Backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/users/me": {
      "get": {
        "operationId": "UsersController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/lookup": {
      "get": {
        "operationId": "UsersController_lookup",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "A phone number in E.164 (`+923001234567`, URL-encode the `+` as %2B) or a username (`@ali_k` or `ali_k`)",
            "schema": {
              "minLength": 3,
              "maxLength": 32,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserLookupDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Find someone to pay by phone or @username. Returns only a masked phone and a short display\nname. Rate-limited more strictly than other reads to slow down enumeration.",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/admin/users": {
      "get": {
        "operationId": "AdminUsersController_find",
        "parameters": [
          {
            "name": "phone",
            "required": true,
            "in": "query",
            "schema": {
              "pattern": "^\\+[1-9]\\d{7,14}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin / users"
        ]
      }
    },
    "/v1/admin/users/{id}": {
      "get": {
        "operationId": "AdminUsersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin / users"
        ]
      }
    },
    "/v1/admin/users/{id}/suspend": {
      "post": {
        "operationId": "AdminUsersController_suspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspendUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin / users"
        ]
      }
    },
    "/v1/admin/users/{id}/reactivate": {
      "post": {
        "operationId": "AdminUsersController_reactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin / users"
        ]
      }
    },
...
    "/v1/admin/ledger/system-accounts": {
      "get": {
        "operationId": "LedgerAdminController_systemAccounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LedgerAccountDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin / ledger"
        ]
      }
    },
    "/v1/admin/ledger/accounts/{id}": {
      "get": {
        "operationId": "LedgerAdminController_account",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerAccountDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin / ledger"
        ]
      }
    },
    "/v1/admin/ledger/entries/{id}": {
      "get": {
        "operationId": "LedgerAdminController_entry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalEntryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin / ledger"
        ]
      }
    },
    "/v1/admin/ledger/entries/{id}/reversals": {
      "post": {
        "operationId": "LedgerAdminController_reverse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Unique key (8-128 chars, [A-Za-z0-9_.:-]) per logical operation. Retries with the same key and payload replay the original response.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReverseEntryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalEntryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin / ledger"
        ]
      }
    },
    "/v1/admin/ledger/integrity": {
      "get": {
        "operationId": "LedgerAdminController_integrity",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrityReportDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin / ledger"
        ]
      }
    },
...
  },
  "info": {
    "title": "PayCore API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Role": {
        "type": "string",
        "enum": [
          "CONSUMER",
          "MERCHANT",
          "ADMIN"
        ]
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "PENDING_VERIFICATION",
          "ACTIVE",
          "SUSPENDED"
        ]
      },
      "KycTier": {
        "type": "string",
        "enum": [
          "TIER_0",
          "TIER_1",
          "TIER_2",
          "TIER_3"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserStatus"
              }
            ]
          },
          "kycTier": {
            "allOf": [
              {
                "$ref": "#/components/schemas/KycTier"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "phoneVerified": {
            "type": "boolean"
          },
          "pinSet": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "role",
          "status",
          "kycTier",
          "id",
          "phone",
          "fullName",
          "phoneVerified",
          "pinSet",
          "createdAt"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "username": {
            "type": "string",
            "description": "3-20 chars of [a-z0-9_]; case-insensitive (stored lower-case) and unique."
          }
        }
      },
      "Currency": {
        "type": "string",
        "enum": [
          "PKR",
          "AED",
          "USD"
        ],
        "description": "Currencies the user can receive in (active wallets)"
      },
      "UserLookupDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "wallets": {
            "type": "array",
            "description": "Currencies the user can receive in (active wallets)",
            "items": {
              "$ref": "#/components/schemas/Currency"
            }
          },
          "userId": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "description": "First name and last initial, e.g. \"Ali K.\""
          },
          "phoneMasked": {
            "type": "string",
            "description": "e.g. \"+92300****567\""
          }
        },
        "required": [
          "username",
          "wallets",
          "userId",
          "displayName",
          "phoneMasked"
        ]
      },
      "SuspendUserDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 5,
            "maxLength": 500
          }
        },
        "required": [
          "reason"
        ]
      },
      "MoneyDto": {
        "type": "object",
        "properties": {
          "currency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "amount": {
            "type": "string",
            "description": "Decimal string in major units, e.g. \"1250.50\"",
            "example": "1250.50"
          },
          "amountMinor": {
            "type": "string",
            "description": "Integer minor units as a string, e.g. \"125050\"",
            "example": "125050"
          }
        },
        "required": [
          "currency",
          "amount",
          "amountMinor"
        ]
      },
...
      "AccountType": {
        "type": "string",
        "enum": [
          "ASSET",
          "LIABILITY",
          "EQUITY",
          "REVENUE",
          "EXPENSE"
        ]
      },
      "NormalBalance": {
        "type": "string",
        "enum": [
          "DEBIT",
          "CREDIT"
        ]
      },
      "AccountOwnerType": {
        "type": "string",
        "enum": [
          "SYSTEM",
          "WALLET",
          "MERCHANT"
        ]
      },
      "LedgerAccountDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountType"
              }
            ]
          },
          "normalBalance": {
            "allOf": [
              {
                "$ref": "#/components/schemas/NormalBalance"
              }
            ]
          },
          "ownerType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountOwnerType"
              }
            ]
          },
          "currency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "balance": {
            "description": "Cached balance on the account's normal side",
            "allOf": [
              {
                "$ref": "#/components/schemas/MoneyDto"
              }
            ]
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "code",
          "type",
          "normalBalance",
          "ownerType",
          "currency",
          "id",
          "name",
          "balance",
          "version"
        ]
      },
      "LedgerAccountDetailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountType"
              }
            ]
          },
          "normalBalance": {
            "allOf": [
              {
                "$ref": "#/components/schemas/NormalBalance"
              }
            ]
          },
          "ownerType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountOwnerType"
              }
            ]
          },
          "currency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "balance": {
            "description": "Cached balance on the account's normal side",
            "allOf": [
              {
                "$ref": "#/components/schemas/MoneyDto"
              }
            ]
          },
          "version": {
            "type": "number"
          },
          "derivedBalance": {
            "description": "Recomputed from postings (source of truth)",
            "allOf": [
              {
                "$ref": "#/components/schemas/MoneyDto"
              }
            ]
          }
        },
        "required": [
          "code",
          "type",
          "normalBalance",
          "ownerType",
          "currency",
          "id",
          "name",
          "balance",
          "version",
          "derivedBalance"
        ]
      },
      "EntryType": {
        "type": "string",
        "enum": [
          "DEPOSIT",
          "WITHDRAWAL",
          "TRANSFER",
          "FX_CONVERSION",
          "REVERSAL",
          "ADJUSTMENT",
          "PAYMENT",
          "REFUND",
          "TOPUP",
          "WITHDRAWAL_HOLD",
          "WITHDRAWAL_PAYOUT",
          "SETTLEMENT",
          "SETTLEMENT_PAYOUT"
        ]
      },
      "PostingDto": {
        "type": "object",
        "properties": {
          "direction": {
            "type": "string",
            "enum": [
              "DEBIT",
              "CREDIT"
            ]
          },
          "currency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "accountId": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "amountMinor": {
            "type": "string"
          },
          "balanceAfter": {
            "type": "string"
          }
        },
        "required": [
          "direction",
          "currency",
          "id",
          "accountId",
          "amount",
          "amountMinor",
          "balanceAfter"
        ]
      },
      "JournalEntryDto": {
        "type": "object",
        "properties": {
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EntryType"
              }
            ]
          },
          "reversalOfId": {
            "type": "string",
            "nullable": true
          },
          "initiatedBy": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "postings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostingDto"
            }
          },
          "id": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "reversalOfId",
          "initiatedBy",
          "metadata",
          "postings",
          "id",
          "description",
          "createdAt"
        ]
      },
      "ReverseEntryDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Why the entry is being reversed (kept on the reversal entry for audit).",
            "minLength": 5,
            "maxLength": 500
          }
        },
        "required": [
          "reason"
        ]
      },
      "UnbalancedEntryDto": {
        "type": "object",
        "properties": {
          "currency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "entryId": {
            "type": "string"
          },
          "net": {
            "type": "string"
          }
        },
        "required": [
          "currency",
          "entryId",
          "net"
        ]
      },
      "NetByCurrencyDto": {
        "type": "object",
        "properties": {
          "currency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "net": {
            "type": "string"
          },
          "postings": {
            "type": "string"
          }
        },
        "required": [
          "currency",
          "net",
          "postings"
        ]
      },
      "BalanceMismatchDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string"
          },
          "cached": {
            "type": "string"
          },
          "derived": {
            "type": "string"
          }
        },
        "required": [
          "accountId",
          "cached",
          "derived"
        ]
      },
      "TrialBalanceDto": {
        "type": "object",
        "properties": {
          "currency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "debitNormal": {
            "type": "string"
          },
          "creditNormal": {
            "type": "string"
          }
        },
        "required": [
          "currency",
          "debitNormal",
          "creditNormal"
        ]
      },
      "IntegrityReportDto": {
        "type": "object",
        "properties": {
          "unbalancedEntries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnbalancedEntryDto"
            }
          },
          "netByCurrency": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NetByCurrencyDto"
            }
          },
          "balanceCacheMismatches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BalanceMismatchDto"
            }
          },
          "trialBalance": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrialBalanceDto"
            }
          },
          "healthy": {
            "type": "boolean"
          }
        },
        "required": [
          "unbalancedEntries",
          "netByCurrency",
          "balanceCacheMismatches",
          "trialBalance",
          "healthy"
        ]
      },
...
    }
  }
}